+++ popup/components/Typeface/index.tsx
@@ -1,47 +1,49 @@
+import { useCallback } from "react"
 import { useStorage } from "@plasmohq/storage/hook"
 import { useSetAtom } from "jotai"
 import { FolderPlus, X } from "phosphor-react"
 
 import { Button } from "./Button"
-
 import { modalOpenAtom, selectedTypefaceSlugAtom } from "~popup/atoms"
+
 import type { ICollection, ITypeface, TypefaceTuple } from "types/typeface"
+import type { SupportedWebsite } from "~types/website"
 
-interface IFavorite {
-  favorite: ITypeface
-}
-
-export const Favorite = ({ favorite }: IFavorite) => {
+export const Typeface = ({ typeface }: { typeface: ITypeface }) => {
   const [favorites, setFavorites] = useStorage<TypefaceTuple[]>("favorites", [])
   const setIsModalOpen = useSetAtom(modalOpenAtom)
   const setSelectedTypeface = useSetAtom(selectedTypefaceSlugAtom)
   const [, setCollections] = useStorage<ICollection[]>("collections", [])
+  const [, setVisibleOrigins] = useStorage<SupportedWebsite[]>("visibleOriginWebsites", [])
 
-  const { origin, family, /** styles, variableAxes **/ slug } = favorite
+  const { origin, family, slug } = typeface
 
-  // const stylesText = `${styles.length} style${styles.length > 1 ? "s" : ""}`
-  // const variableAxesText = variableAxes > 0 ? `(variable - ${variableAxes} axes)` : ""
-
-  const openModal = () => {
+  const openModal = useCallback(() => {
     setIsModalOpen("collection-assignment")
-    setSelectedTypeface(favorite.slug)
-  }
+    setSelectedTypeface(typeface.slug)
+  }, [typeface])
 
-  const handleRemoveTypeface = () => {
+  const handleRemoveTypeface = useCallback(() => {
     const favoritesMap = new Map(favorites)
     favoritesMap.delete(slug)
     setFavorites(Array.from(favoritesMap))
     setCollections((prev) =>
       prev.map((collection) => {
-        if (!collection.typefaces.includes(favorite.slug)) return collection
+        if (!collection.typefaces.includes(typeface.slug)) return collection
         return {
           ...collection,
-          typefaces: collection.typefaces.filter(
-            (font) => font !== favorite.slug
-          )
+          typefaces: collection.typefaces.filter((font) => font !== typeface.slug)
         }
       })
     )
+    setVisibleOrigins((prev) => {
+      const remainingFontsWithSameOrigin = favorites.filter(
+        (fav) => fav[1].slug !== slug && fav[1].origin.name === origin.name
+      )
+      if (remainingFontsWithSameOrigin.length === 0) {
+        return prev.filter((prevOrigin) => prevOrigin !== origin.name)
+      }
+    })
 
     // Send message to the content_script that a font has been removed from wishlist.
     // The content_script uses this message to change the state of the add/remove wishlist button.
@@ -51,7 +53,7 @@ export const Favorite = ({ favorite }: IFavorite) => {
         font: slug
       })
     })
-  }
+  }, [favorites, slug])
 
   return (
     <div className="group flex items-center justify-between border-b border-greyscale-200 px-4 py-3 ">
@@ -62,9 +64,6 @@ export const Favorite = ({ favorite }: IFavorite) => {
           className="mb-[2px] block text-lg font-medium no-underline [line-height:1.2] hover:cursor-ne-resize hover:underline">
           {family}
         </a>
-        {/* <p className="text-sm text-greyscale-600 opacity-90">
-          {stylesText} {variableAxesText}
-        </p> */}
         <p className="text-sm text-greyscale-600 opacity-90">{origin.name}</p>
       </div>
       <div className="flex gap-2 opacity-0 group-hover:opacity-100">