+++ tailwind.config.js
@@ -5,9 +5,26 @@ module.exports = {
   theme: {
     extend: {
       fontFamily: {
-        poppins: ['Poppins', 'sans-serif']
+        poppins: ['Poppins', 'sans-serif'],
+        inherit: ['inherit']
       },
       colors: {
+        gf: {
+          dark: {
+            primary: '#8ab4f8',
+            secondary: '#d2e3fc'
+          },
+          light: {
+            primary: '#1a73e8',
+            secondary: '#174ea6'
+          }
+        },
+        adobe: {
+          light: {
+            primary: '#707070',
+            secondary: '#e1e1e1'
+          }
+        },
         greyscale: {
           100: '#fbfbfe',
           200: '#dadde6',